service: Support owning clock.
Current design of DataStructureService enforces its owner to pass in
clock, this allows the unit test to use a dummy clock to test TTL. On
the other hand, it gives server a burden to manage a system clock.

This patch updated DataStructureService to manage its own clock in
case of nullptr passed in as clock. Now the server no longer needs to
create the clock for DSS.