io: Add buffer ring agnostic Recv to Connection
Currently, the user of Connection issues buffer ring version or normal
version of Recv explicitly, which is suboptimal: The user should not
need to know whether buffer ring is used since it makes no sense to the
application layer.

This patch enabled Connection to provide buffer ring agnostic Recv: a
Recv that takes a pointer to buffer. When used as normal, it performs
normal version of recv that copies data into buffer's sink. To use it
with buffer ring, the user can set Connection via 'SetUseRingBuf'. It
will fill 'buffer' with view over ring buffer.